Grid GUI: cleanup
